properly fixed the excitation/source scaling in the domain, added beam steering example
When introducing the source term in the Westervelt equation the source term has SI units kg/(m s^4) which is the second time derivate of the source term. Therefore, also in the Helmholtz system that we obtain by projection into the frequency domain, under the assumption that the source is a pressure in Pa, we have to project the second time derivate!